requirements.txt
mido[ports-rtmidi]
pillow
